slowly s/handle/store
